fix stop shell

+++ scripts/stop.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-test -f RUNNING_PID && kill `cat RUNNING_PID` && sleep 5;
-rm RUNNING_PID;
+test -f RUNNING_PID && kill `cat RUNNING_PID` && sleep 5 && rm RUNNING_PID;
+
